requirements.txt
Django==3.1.3
Pillow==10.0.0
jinja2==3.0
gunicorn==20.1.0
contextvars
typing-extensions==4.2.0
aiohttp
click==8.0.4
wheel==0.41.1
multidict==4.5
psycopg2-binary==2.9.7